+++ Competencias/Roboliga_2020/Titan/Equipo 1/Codigo/Corto plazo/Para hacer en esta semana.md
