+++ Competencias/Roboliga_2020/Titan/Equipo 1/Codigo/Largo plazo/GlobalSensorDetection.py
