+++ Competencias/Roboliga_2020/Titan/Equipo 1/Codigo/Largo plazo/Para hacer en esta semana.md
