+++ Competencias/Roboliga_2020/Titan/Equipo 1/Codigo/OpenCV/Para hacer en esta semana.md
